pyproject.toml
[tool.poetry]
name = "pymosaic-4k"
version = "0.1.0"
description = "Create 4K mosaics"
authors = ["Quentin Barbarat <quentin.barbarat@epita.fr>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.3.0"
typer = "^0.3.2"
fastapi = "^0.66.0"
python-multipart = "^0.0.5"
aiofiles = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"
flake8 = "^3.9.2"
isort = "^5.9.1"
pre-commit = "^2.13.0"
uvicorn = {extras = ["standard"], version = "^0.14.0"}

[tool.poetry.scripts]
mosaic = "pymosaic_4k.cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
